Continued work on hand-raising robot (#4)
With this version, pushing the button cycles through 3 behaviors: raising the hand, wiggling the hand, and lowering it again (and if you push the button more, it raises again, wiggles again, etc.). We also made the robot a context manager so it can clean itself up properly. 

Next steps (which I personally will need help on; I don't know how to make progress on any of these):
- make some way for the external world to interact with it. Zoom API? Something else? We can't have the remote people come to the office to push the button. :grin: 
- get an actual hand-like thing to put on the end of the servo so someone will notice when you raise the hand


* start of movement code, doesn't work yet

* debugging

* make the robot a context manager

* make main an executable

* whitespace tweaks