Count people, wiggle hand when idle (#5)
The class now stores how many people have their hand raised: you can increment or decrement this value, or set it to some specific value (which might be unused in the final product, but will help with a PoC before we integrate with Zoom). When the hand has been raised for sufficiently long, it wiggles (or is supposed to, see below).

The current main function now uses the button to toggle between 1 person raising and lowering the hand (so, it doesn't actually demonstrate that we do the right thing multiple people yet). I imagine the next PR after this will be to add a server that takes in requests to raise and lower the hand, at which point we can get rid of the button and the busy loop, and demonstrate support for multiple people (and also enable remote folks to play with the code, since you no longer need to push the button).


* start on a thread to wiggle the hand on inactivity

* names of non-public functions should start with an underscore

* start and stop the thread as relevant when setting the count, too

* The condition variable should have its own mutex

* simplify main function: button just raises and lowers, since wiggling is idle now

* non-public things start with underscores

* Make self._thread None when the thread isn't supposed to be running

* avoid race condition: don't start the thread until the hand is raised

* fix syntax error

* the robot is not a private member (yet)

* debugging

* avoid race condition: always stop the thread before lowering the hand

* tweak: make the code look symmetric

* avoid deadlock: only acquire the condition variable when you're using it

* lock the mutex whenever you interact with viam. Maybe not needed?

* use coroutines instead of threads, needs debugging info removed when ready

* add docstrings

* move _wiggle_on_inactivity up near _start_wiggler and _stop_wiggler

* debugging: await coroutines

* debugging

* more debugging

* try an Event instead of a condition variable

* things work, so cleaning up

* only use the mutex when we mutate state

* use asyncio.sleep instead of time.sleep

* tweak comments

* remove more debugging logic

* cancel the task, rather than setting an event to tell the task to cancel itself

* Add another docstring

* debugging