Refactoring: use contextlib and other tweaks (#8)
No behavioral changes are intended. 

Tried out on Nicole's desk: things work as intended, except that there are sometimes weird pauses in the middle of wiggling (e.g., instead of getting 3 wiggles, we get 1 wiggle, then a pause, and then 2 wiggles). I doubt that's caused by the hardware, and I don't see how it could be caused by this software, but I also feel uncomfortable just blaming gremlins in the wifi. It would be neat to try running everything on a single board and seeing if that gets rid of the delays...


* rename Robot._robot -> Robot._client

* add docstring

* use contextlib instead of raw __aenter__ functions

* split the context manager into helper functions

* Pass in the credentials to the context manager. Hooray, pure functions!

* add/tweak docstrings

* rename pi -> board. It might not always be a pi.

* Split the Robot class out to a separate file

* tweak more docstrings

* Move Robot.create out of the class, now that it's in a robot-specific file

* tweaks

* alphabetize imports

* the class file doesn't need to be executable

* debugging

* fix docstring

* better encapsulation: the robot only cares about the servo, and the board comes straight from the context manager

* wrap long lines