Fix Chromium hang, refactor slightly (#31)
Without this change, Chromium hangs for me on Linux right now, and with this change, things work normally. I made the change suggested by StackOverflow, and then the options-related stuff was long enough that I split it out to a separate helper function. Then, the browser-initialization stuff was long enough I split it all out to a separate file.

I hope this still works on a Mac; I'm unable to test that myself...


* get Chromium to work even if another Chromium window is already opened

* move Chrome options to a helper function

* Move browser stuff to a separate file from Zoom stuff